Add comments, fan-out feed, and WS notifications; harden for deployment
- Full comments feature (rate-limited route + threaded UI), replacing the stub
- Fan-out-on-write feed via Redis sorted sets (feedline:<userId>) instead of
  a read-time join, with follow-time backfill and unfollow-time purge
- Authenticated WebSocket (/ws?token=<jwt>) with per-user targeted delivery
  (wsHub.sendToUser) for new-follower notices, routed through Redis pub/sub
  so it stays correct across more than one backend replica
- Production hardening: per-IP auth rate limiting, sanitized error
  responses, localhost-only datastore ports, graceful shutdown, and schema
  migrations (node-pg-migrate) replacing manual DB changes
- Kubernetes manifests for local kind practice and a Playwright E2E suite
  covering auth, feed, follow, like/comment, and growth-toast regressions

+++ backend/src/routes/projects.js
@@ -79,8 +79,8 @@ projectsRouter.get("/:ownerId/:slug", requireAuth, ah(async (req, res) => {
   const postsResult = await pool.query(
     `SELECT p.id, p.user_id, p.post_type, p.content, p.media_url, p.created_at,
             u.username, u.display_name, u.avatar_url,
-            (SELECT count(*) FROM likes WHERE post_id = p.id) AS like_count,
-            (SELECT count(*) FROM comments WHERE post_id = p.id) AS comment_count,
+            (SELECT count(*)::int FROM likes WHERE post_id = p.id) AS like_count,
+            (SELECT count(*)::int FROM comments WHERE post_id = p.id) AS comment_count,
             EXISTS(SELECT 1 FROM likes WHERE post_id = p.id AND user_id = $2) AS liked_by_me
      FROM posts p
      JOIN users u ON u.id = p.user_id
